# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend Developer Node.js Study Lab API
  version: 1.1.0
  description: >-
    Complete OpenAPI reference for the backend study cases.
    Includes request body schemas, request examples, and response examples
    for both success and common error outcomes.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /cases/05-timescale/setup:
    post:
      tags: [Data and Platform]
      summary: Setup Timescale extension and hypertable
      operationId: setupCase05Timescale
      responses:
        "200":
          description: Hypertable created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 05-timescale
                    message: hypertable ready
        "500":
          description: Setup failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                setupFailed:
                  value:
                    error: Timescale setup failed
...
  /cases/05-timescale/query:
    get:
      tags: [Data and Platform]
      summary: Query aggregated metrics
      operationId: queryCase05Metrics
      parameters:
        - in: query
          name: service
          required: false
          schema:
            type: string
            default: api
          description: Service name filter
          example: api
        - in: query
          name: minutes
          required: false
          schema:
            type: integer
            default: 60
            minimum: 1
          description: Sliding time window in minutes
          example: 60
      responses:
        "200":
          description: Aggregated metric buckets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 05-timescale
                    service: api
                    minutes: 60
                    data:
                      - bucket: 2026-04-11T09:10:00.000Z
                        avg_value: 121.2
                      - bucket: 2026-04-11T09:05:00.000Z
                        avg_value: 117.4
        "500":
          description: Query failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                queryFailed:
                  value:
                    error: Timescale query failed
  /cases/06-hasura/action/greet:
    post:
      tags: [Data and Platform]
      summary: Hasura action webhook sample
      operationId: runCase06HasuraAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              hasuraAction:
                value:
                  session_variables:
                    x-hasura-role: developer
                  input:
                    name: backend learner
      responses:
        "200":
          description: Action response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 06-hasura
                    message: Hello backend learner, request accepted for role developer
  /cases/06-hasura/event/order-created:
    post:
      tags: [Data and Platform]
      summary: Hasura event-trigger webhook sample
      operationId: runCase06HasuraEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              hasuraEvent:
                value:
                  event:
                    id: 9f0c7647-798e-42f5-a56a-7f80bc10d2dc
                    data:
                      new:
                        id: ord_123
                        amount: 99
                      old: null
      responses:
        "200":
          description: Event callback ack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 06-hasura
                    received: true
                    eventId: 9f0c7647-798e-42f5-a56a-7f80bc10d2dc
                    table: insert
...
  /cases/08-microservices/setup:
    post:
      tags: [Distributed Systems]
      summary: Setup orders and outbox tables
      operationId: setupCase08
      responses:
        "200":
          description: Tables ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 08-microservices
                    message: service_orders and outbox_events are ready
        "500":
          description: Setup failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                setupFailed:
                  value:
                    error: Setup failed
  /cases/08-microservices/orders:
    post:
      tags: [Distributed Systems]
      summary: Create order and outbox event in one transaction
      operationId: createCase08Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId:
                  type: string
                amount:
                  type: number
                  exclusiveMinimum: 0
              required: [customerId, amount]
            examples:
              createOrder:
                value:
                  customerId: user-1
                  amount: 299
      responses:
        "201":
          description: Order accepted and event queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 08-microservices
                    orderId: 96e356af-e077-4a5e-9abf-8fdb8f6ef7cc
                    outboxEventId: 72c49fe9-7d63-49bc-8ca9-bdcbd12fcf75
        "400":
          description: Invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidAmount:
                  value:
                    error: amount must be greater than 0
        "500":
          description: Order create failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                createFailed:
                  value:
                    error: Order create failed
  /cases/08-microservices/process-outbox:
    post:
      tags: [Distributed Systems]
      summary: Process next pending outbox event
      operationId: processCase08Outbox
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                processed:
                  value:
                    case: 08-microservices
                    processed: true
                    eventId: 72c49fe9-7d63-49bc-8ca9-bdcbd12fcf75
                emptyQueue:
                  value:
                    case: 08-microservices
                    processed: false
        "500":
          description: Outbox processing failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                processFailed:
                  value:
                    error: Outbox processing failed
  /cases/08-microservices/state:
    get:
      tags: [Distributed Systems]
      summary: Read order and outbox state
      operationId: getCase08State
      responses:
        "200":
          description: Current order and outbox records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 08-microservices
                    orders:
                      - id: 96e356af-e077-4a5e-9abf-8fdb8f6ef7cc
                        customerId: user-1
                        amount: 299
                        status: processed
                        createdAt: 2026-04-11T09:22:00.000Z
                    outbox:
                      - id: 72c49fe9-7d63-49bc-8ca9-bdcbd12fcf75
                        eventType: order.created
                        payload:
                          id: 96e356af-e077-4a5e-9abf-8fdb8f6ef7cc
                          customerId: user-1
                          amount: 299
                          status: created
                        published: true
                        createdAt: 2026-04-11T09:22:00.000Z
                        publishedAt: 2026-04-11T09:22:01.000Z
        "500":
          description: State query failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                stateFailed:
                  value:
                    error: State query failed
...
  /cases/09-queue-worker/setup:
    post:
      tags: [Distributed Systems]
      summary: Setup async jobs table
      operationId: setupCase09
      responses:
        "200":
          description: Queue table ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 09-queue-worker
                    message: async_jobs table ready
        "500":
          description: Setup failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                setupFailed:
                  value:
                    error: Setup failed
  /cases/09-queue-worker/enqueue:
    post:
      tags: [Distributed Systems]
      summary: Enqueue a job for worker processing
      operationId: enqueueCase09Job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueEnqueueRequest"
            examples:
              enqueue:
                value:
                  type: send-email
                  payload:
                    to: candidate@example.com
                  maxAttempts: 3
      responses:
        "201":
          description: Job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 09-queue-worker
                    job:
                      id: 9f6ac3b9-11c6-4e9a-ad95-2a2cf06dfad0
                      type: send-email
                      payload:
                        to: candidate@example.com
                      attempts: 0
                      maxAttempts: 3
                      status: pending
                      lastError: null
                      availableAt: 2026-04-11T09:26:10.000Z
                      createdAt: 2026-04-11T09:26:10.000Z
                      processedAt: null
        "500":
          description: Enqueue failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                enqueueFailed:
                  value:
                    error: Enqueue failed
  /cases/09-queue-worker/process:
    post:
      tags: [Distributed Systems]
      summary: Process the next pending job
      operationId: processCase09Job
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                processed:
                  value:
                    case: 09-queue-worker
                    processed:
                      id: 9f6ac3b9-11c6-4e9a-ad95-2a2cf06dfad0
                      type: send-email
                      payload:
                        to: candidate@example.com
                      attempts: 1
                      maxAttempts: 3
                      status: completed
                      lastError: null
                      availableAt: 2026-04-11T09:26:10.000Z
                      createdAt: 2026-04-11T09:26:10.000Z
                      processedAt: 2026-04-11T09:26:11.100Z
                empty:
                  value:
                    case: 09-queue-worker
                    message: No pending jobs
        "500":
          description: Processing failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                processFailed:
                  value:
                    error: Process failed
  /cases/09-queue-worker/stats:
    get:
      tags: [Distributed Systems]
      summary: Read pending/processing/completed/dead-letter buckets
      operationId: getCase09Stats
      responses:
        "200":
          description: Queue state buckets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 09-queue-worker
                    pending: []
                    processing: []
                    completed:
                      - id: 9f6ac3b9-11c6-4e9a-ad95-2a2cf06dfad0
                        type: send-email
                        payload:
                          to: candidate@example.com
                        attempts: 1
                        maxAttempts: 3
                        status: completed
                        lastError: null
                        availableAt: 2026-04-11T09:26:10.000Z
                        createdAt: 2026-04-11T09:26:10.000Z
                        processedAt: 2026-04-11T09:26:11.100Z
                    deadLetter: []
        "500":
          description: Stats query failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                statsFailed:
                  value:
                    error: Stats query failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      required: [error]
...
    QueueEnqueueRequest:
      type: object
      properties:
        type:
          type: string
        payload:
          type: object
          additionalProperties: true
        maxAttempts:
          type: integer
          minimum: 1
          default: 3
      required: [type]
...